% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUIDEseq-package.R
\docType{data}
\name{peaks.gr}
\alias{peaks.gr}
\title{example cleavage sites}
\format{
\describe{GRanges with count (peak height), bg (local background),
SNratio (signal noise ratio), p-value, and option adjusted p-value }
}
\source{
http://trace.ncbi.nlm.nih.gov/Traces/sra/?run=SRR1695644
}
\value{
\item{peaks.gr}{GRanges with count (peak height), bg (local
background), SNratio (signal noise ratio), p-value, and option adjusted
p-value }
}
\description{
An example data set containing cleavage sites (peaks) from getPeaks
}
\examples{

    data(peaks.gr)
    names(peaks.gr)
    peaks.gr

}
\keyword{datasets}
