% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTracks.R
\name{plotTracks}
\alias{plotTracks}
\title{Plot offtargets along all chromosomes with one track per chromosome}
\usage{
plotTracks(
  offTargetFile,
  sep = "\\t",
  header = TRUE,
  chromosome.order = paste0("chr", c(1:22, "X", "Y")),
  xlab = "Chromosome Size (bp)",
  ylab = "Peak Score",
  score.col = c("peak_score", "total.mismatch.bulge", "predicted_cleavage_score"),
  transformation = c("log10", "none"),
  title = "",
  axis.title.size = 12,
  axis.label.size = 8,
  strip.text.y.size = 9,
  off.target.line.size = 0.6,
  on.target.line.size = 1,
  on.target.score = 1,
  on.target.color = "red",
  off.target.color = "black",
  strip.text.y.angle = 0,
  scale.grid = c("free", "free_y", "free_x", "fixed")
)
}
\arguments{
\item{offTargetFile}{The file path containing off-targets generated
from GUIDEseqAnalysis}

\item{sep}{The separator in the file, default to tab-delimited}

\item{header}{Indicates whether the input file contains a header,
default to TRUE}

\item{chromosome.order}{The chromosome order to plot from top to bottom}

\item{xlab}{The x-asix label, default to Chromosome Size (bp)}

\item{ylab}{The y-asix label, default to Peak Score. Change it to
be consistent with the score.col}

\item{score.col}{The column used as y values in the plot. Available choices
are peak_score, total.mismatch.bulge, and predicted_cleavage_score.}

\item{transformation}{Indicates whether plot the y-value in log10 scale or
in the original scale. When scale.col is set to total.mismatch.bulge, the
data will be plotted in the original scale.}

\item{title}{The figure title, default to none.}

\item{axis.title.size}{The font size for the axis labels, default to 12}

\item{axis.label.size}{The font size for the tick labels, default to 8}

\item{strip.text.y.size}{The font size for the strip labels, default to 9}

\item{off.target.line.size}{The line size to depict the off-targets, default
to 0.6}

\item{on.target.line.size}{The line size to depict the on-targets, default
to 1}

\item{on.target.score}{The score for the on-target, default to 1 for CFD
scoring system. This is the maximum score in the chosen scoring system.
Change it accordingly if different off-target scoring system is used.}

\item{on.target.color}{The line color to depict the on-targets, default
to red}

\item{off.target.color}{The line color to depict the off-targets, default
to black}

\item{strip.text.y.angle}{The angel for the y strip text, default to 0.
Set it to 45 if angled representation is desired}

\item{scale.grid}{Used to set the scales in facet_grid, default to free,
meaning that scales vary across both rows and columns.
Other options are fixed, free_x, and free_y meaning that scales shared
across all facets, vary across rows, and vary across columns, respectively.
For details, please type ?ggplot2::facet_grid}
}
\value{
a ggplot object
}
\description{
Plot offtargets along all chromosomes with one track per chromosome
}
\examples{
if (interactive())
{
   offTargetFilePath <- system.file("extdata/forVisualization",
      "offTargetsInPeakRegions.xls",
       package = "GUIDEseq")
  fig1 <- plotTracks(offTargetFile = offTargetFilePath)
  fig1
  fig2 <- plotTracks(offTargetFile = offTargetFilePath,
    score.col = "total.mismatch.bulge",
    ylab = "Total Number of Mismatches and Bulges")
  fig2
  fig3 <- plotTracks(offTargetFile = offTargetFilePath,
     score.col = "predicted_cleavage_score",
     ylab = "CFD Score",
     scale.grid = "fixed",
     transformation = "none")
 fig3
}
}
\author{
Lihua Julie Zhu
}
