% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatmapOfftargets.R
\name{plotHeatmapOfftargets}
\alias{plotHeatmapOfftargets}
\title{Plot offtargets as heatmap}
\usage{
plotHeatmapOfftargets(
  mergedOfftargets,
  min.detection.rate = 0.2,
  font.size = 12,
  on.target.predicted.score = 1,
  MR.normalization = c("sequence.depth", "peak.score"),
  top.bottom.height.ratio = 3,
  dot.distance.breaks = c(5, 10, 20, 40, 60),
  dot.distance.scaling.factor = c(0.4, 0.6, 0.8, 1.2, 2),
  bottom.start.offset = 8,
  color.low = "white",
  color.high = "blue",
  sample.names
)
}
\arguments{
\item{mergedOfftargets}{a data frame from running combineOfftargets}

\item{min.detection.rate}{minimum indel detection rate to be included
in the heatmap}

\item{font.size}{font size for x labels and numbers along the y-axis.}

\item{on.target.predicted.score}{Default to 1 for the CFDscore
scoring method. Set it to 100 for the Hsu-Zhang
scoring method.}

\item{MR.normalization}{Default to peak.score, i.e., using the sum of
peak.score for each sample in the input file to calculate mutation rate.
Set it to sequence.depth to use the sequencing depth for each sample
in the input file to calculate mutation rate.}

\item{top.bottom.height.ratio}{the ratio of the height of top panel
vs that of the bottom panel.}

\item{dot.distance.breaks}{a numeric vector for specifying
the minimum number of rows in each panel to use the
the corresponding distance in dot.distance.scaling.factor
betwen consecutive dots along the y-axis.
In the default setting, dot.distance.breaks and
dot.distance.scaling.factor are
set to c(5, 10, 20, 40, 60) and c(0.4, 0.6, 0.8, 1.2, 2)
respectively, which means that if the number of rows in each panel
is greater than or equal to 60, 40-59, 20-39, 10-19, 5-9, and less
than 5,then the disance between consecutive dots will
be plotted 2, 1.2, 0.8, 0,6, 0.4, and 0.2 (half of 0.4) units away
in y-axis respectively.}

\item{dot.distance.scaling.factor}{a numeric vector for specifing
the distance between two consecutive dots.
See dot.distance.breaks for more information.}

\item{bottom.start.offset}{Default to 2, means that place the
top number in the bottom panel 2 units below the top border. Increase
the value will move the number away from the top border.}

\item{color.low}{The color used to represent the lowest indel rate,
default to white}

\item{color.high}{The color used to represen the highest indel rate
the itermediate indel rates will be colored using the color between
color.low and color.high. Default to blue.}

\item{sample.names}{Optional sample Names used to label the x-axis. If
not provided, x-axis will be labeled using the sample names provided
in the GUIDEseqAnalysis step.}
}
\value{
a ggplot object
}
\description{
Plot offtargets as heatmap
}
\examples{
if(interactive())
{
figs = plotHeatmapOfftargets(mergedOfftargets,
    min.detection.rate = 2,
    MR.normalization = "peak.score",
    top.bottom.height.ratio = 6,
    bottom.start.offset = 8,
    dot.distance.scaling.factor = c(0.2,0.2,0.6,0.8, 1.2))
figs[[1]]/figs[[2]] +
     plot_layout(heights = unit(c(2.6,1),
                        c('null', 'null')))

plotHeatmapOfftargets(mergedOfftargets,
    min.detection.rate = 0,
    MR.normalization = "peak.score",
   top.bottom.height.ratio = 6,
   bottom.start.offset = 8,
   dot.distance.scaling.factor = c(0.2,0.4,0.6,0.8, 4))


figs = plotHeatmapOfftargets(mergedOfftargets,
    min.detection.rate = 0.1,
    MR.normalization = "peak.score",
    top.bottom.height.ratio = 5,
    bottom.start.offset = 8,
    dot.distance.scaling.factor = c(0.2,0.2,0.6,0.8, 1.2))
figs[[1]]/figs[[2]] +
     plot_layout(heights = unit(c(3,1),
                 c('null', 'null')))

 figs = plotHeatmapOfftargets(mergedOfftargets,
     min.detection.rate = 0.5,
     MR.normalization = "peak.score",
    top.bottom.height.ratio = 12,
    bottom.start.offset = 12,
    dot.distance.scaling.factor = c(0.2,0.2,0.6,0.8, 1.2),
    sample.names = c("Control", "Treated"))
figs[[1]]/figs[[2]] +
   plot_layout(heights = unit(c(2,1),
                           c('null', 'null')))
}
}
\author{
Lihua Julie Zhu

Lihua Julie Zhu
}
